requirements.txt
websockets>=12.0
requests>=2.31
pyyaml>=6.0
